+++ lib/widgets/markdown_field.dart
@@ -122,8 +122,7 @@ class MarkdownField extends StatelessWidget {
     return Padding(
       padding: padding,
       child: TextField(
-        key: key,
-        // const ValueKey<String>("markdown_editor_plus"),
+        // key: const ValueKey<String>("markdown_editor_plus"),
         maxLines: expands ? null : maxLines,
         minLines: expands ? null : minLines,
         expands: expands,